Start of simple Lua bots
