+++ dlls/ff/ff_bot_debug.h
@@ -1,0 +1,19 @@
+//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//
+// Purpose: 
+//
+// $NoKeywords: $
+//=============================================================================//
+
+#ifndef FF_BOT_DEBUG_H
+#define FF_BOT_DEBUG_H
+#ifdef _WIN32
+#pragma once
+#endif
+
+#ifdef _DEBUG
+
+void Bot_RunAll();
+
+#endif // _DEBUG
+#endif // FF_BOT_DEBUG_H
